video_subtitles/requirements.txt
openai-whisper
pydub
deep-translator
tqdm
torch
